Title: 삼각형 만들기, Time: 130ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240102/삼각형 만들기/create-triangle.java
@@ -23,6 +23,6 @@ public class Main {
             }
         }
 
-        System.out.print((int) max * 2);
+        System.out.print(max * 2);
     }
 }
